+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support standalone Tools to run without `CompilationUnit`: PR [#326](https://github.com/nowarp/misti/pull/326)
+- Logger: Detector contexts: PR [#327](https://github.com/nowarp/misti/pull/327)
 
 ## [0.7.1] - 2025-03-05